.history/backend/requirements_20250711130921.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Pydantic & Validation
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# AI & Machine Learning (versões compatíveis)
openai>=1.6.1
langchain
langchain-openai
ollama
numpy
pandas

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Notifications
twilio
sendgrid
boto3

# Image Processing
pillow
python-magic

# Testing
pytest
pytest-asyncio
httpx
pytest-cov

# Utilities
python-slugify
pyqrcode
pypng
reportlab
openpyxl

# Development
black
isort
flake8

# Monitoring
sentry-sdk[fastapi] 